+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ### Added:
 - Added respawn option
 ### Fixed:
-- 
+- Fixed dungeon on /leave
 ### Changed:
 -